+++ src/container/notFound/NotFoundPicture/index.style.tsx
