+++ src/container/notFound/NotFoundText/index.style.tsx
